[GridNG] Apply justify-content, alignment-content.
As above. This patch implements the core content alignment logic. E.g.
determining the offset at which the tracks should start, and the space
between them (gutter-size - this include the grid-gap if any).

Creates a new struct "TrackAlignmentGeometry" which contains these
two values. Effectively the process is now:
GridGap -> TrackAlignmentGeometry -> SetOffsets.

The most significant difference is that TrackAlignmentGeometry includes
the grid-gap, and the BorderScrollPadding start offset.

Bug: 1045599
Change-Id: Idb4d550f0d8b2aca306945a46d41b5a30a20bca4